Done System Wake.

+++ Assets/Source/Script/Data/ActorStat.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class ActorStat : ScriptableObject
 {
-    public int Hp;
+    public float Hp;
     public float MoveSpeed;
-    public int Damage;
+    public float Damage;
 }
